moved git utils and use gitpython
